+++ mailet/crypto/src/main/java/org/apache/james/transport/mailets/SMIMESign.java
@@ -19,7 +19,7 @@
 
 
 
-package org.apache.james.transport.mailet;
+package org.apache.james.transport.mailets;
 
 import org.apache.mailet.Mail;
 
@@ -31,18 +31,12 @@ import javax.mail.internet.MimeMultipart;
 import java.io.IOException;
 
 /**
- * <p>Puts a <I>server-side</I> signature on a message.
- * It is a concrete subclass of {@link AbstractSign}, with very few modifications to it.</p>
- * <p>A text file with an explanation text is attached to the original message,
- * and the resulting message with all its attachments is signed.
- * The resulting appearence of the message is almost unchanged: only an extra attachment
- * and the signature are added.</p>
- *<p>The kind of signuture depends on the value of the &lt;keyHolderClass&gt; init parameter.
+ * <p>Puts a <I>server-side</I> SMIME signature on a message.
+ * It is a concrete subclass of {@link Sign}, with very few modifications to it,
+ * to specialize for SMIME.</p>
  *
  *  <P>Handles the following init parameters (will comment only the differences from {@link AbstractSign}):</P>
  * <ul>
- * <li>&lt;keyHolderClass&gt;: Sets the class of the KeyHolder object that will handle the cryptography functions,
- * for example org.apache.james.security.SMIMEKeyHolder for SMIME.</li>
  * <li>&lt;debug&gt;.</li>
  * <li>&lt;keyStoreFileName&gt;.</li>
  * <li>&lt;keyStorePassword&gt;.</li>
@@ -56,9 +50,9 @@ import java.io.IOException;
  * displaying also all the headers of the original message (see {@link #getExplanationText}).</li>
  * </ul>
  * @version CVS $Revision$ $Date$
- * @since 2.2.1
+ * @since 2.3.0
  */
-public class Sign extends AbstractSign {
+public class SMIMESign extends Sign {
     
     /**
      * Return a string describing this mailet.
@@ -66,7 +60,7 @@ public class Sign extends AbstractSign {
      * @return a string describing this mailet
      */
     public String getMailetInfo() {
-        return "Signature Mailet";
+        return "SMIME Signature Mailet";
     }
     
     /**
@@ -74,7 +68,6 @@ public class Sign extends AbstractSign {
      */
     protected  String[] getAllowedInitParameters() {
         return new String[]{
-            "keyHolderClass",
             "debug",
             "keyStoreFileName",
             "keyStorePassword",
@@ -127,6 +120,22 @@ public class Sign extends AbstractSign {
         }
         
         return explanationText;
+    }
+    
+    /**
+     * Initializer for property keyHolderClass.
+     * Hardcodes it to {@link org.apache.james.transport.SMIMEKeyHolder}.
+     */
+    protected void initKeyHolderClass() throws MessagingException {
+        String keyHolderClassName = "org.apache.james.security.SMIMEKeyHolder";
+        try {
+            setKeyHolderClass(Class.forName(keyHolderClassName));
+        } catch (ClassNotFoundException cnfe) {
+            throw new MessagingException(keyHolderClassName + "does not exist.");
+        }
+        if (isDebug()) {
+            log("keyHolderClass: " + getKeyHolderClass());
+        }
     }
     
     /**
